cicd: update restart-app .sh

+++ restart-app.sh
@@ -1,4 +1,8 @@
 PID=$(ps | grep app.py | grep -v grep | awk '{print $1}')
-echo $PID
-kill $PID
+if [ -z "$PID" ]; then
+    echo "No process to kill"
+else
+    echo "Killing process $PID"
+    kill $PID
+echo "Starting app.py..."
 nohup python app.py &
